Use dummy_verify fixture for running database auth unit tests
